scrapers/car_gurus/requirements.txt
async-generator==1.10
attrs==21.4.0
boto3==1.20.28
botocore==1.23.28
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
h11==0.12.0
idna==3.3
jmespath==0.10.0
numpy==1.21.5
outcome==1.1.0
pandas==1.3.5
psycopg2==2.9.2
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.3
urllib3==1.26.7
websockets==10.1
wsproto==1.0.0
